chore: Implement a rebate page that supports multiple languages
