+++ apps/web/src/language-data/AbpUiNavigation/resources/en.json
@@ -37,5 +37,6 @@
   "Identity:claimType": "Claim Types",
   "Identity:securityLogs": "Security Logs",
   "SaasManagement": "SaaS Management",
-  "Investmensts": "Investmensts"
+  "Investmensts": "Investmensts",
+  "Management": "Management"
 }
