+++ apps/web/src/language-data/AbpUiNavigation/resources/tr.json
@@ -37,5 +37,6 @@
   "Identity:claimType": "Talep Türleri",
   "Identity:securityLogs": "Güvenlik Kayıtları",
   "SaasManagement": "SaaS Yönetimi",
-  "Investmensts": "Yatırımlar"
+  "Investmensts": "Yatırımlar",
+  "Management": "Yönetim"
 }
